requirements.txt
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
bokeh==2.3.3
certifi==2021.5.30
cffi==1.14.5
channels==3.0.3
chardet==4.0.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.4
freeopcua==0.90.6
gevent==21.1.2
greenlet==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.0
opcua==0.98.13
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
scipy==1.7.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
tornado==6.1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
websocket-client==1.1.0
websockets==9.1
zope.event==4.5.0
zope.interface==5.4.0
